Add files via upload
I (Ilia Blinderman) began compiling a three-part guide, intended for beginners with next to no technical knowledge, on how to create the sorts of data-driven, visual stories that we publish at The Pudding. This was going to consist of introductions and surveys of data analysis, design, and writing, and while I thought that the first two sections would be all that our readers really wanted to see, we've consistently received word that there remains a substantial interest in the writing guide. I'd tried to compile this third section, which had maintained its "On Writing" heading in my mind, with little success, until I realized that the impasse lay in the fact that I was fixated on precisely that: writing. Compiling resources on how to write within the form of a visual essay while omitting the discussion of storytelling, in both visual and written terms, as well as the importance of story structure, is of next to no help. Thus, I decided to retroactively make good and compile a primer on storytelling with data.